H_W 5

+++ src/edu/Hillel/JavaForQA/Homework/Lesson_4/Task_2_1/MainAutomorph.java
@@ -19,10 +19,10 @@ public class MainAutomorph {
 
         int num = sc.nextInt();
 
-        int sq_num = num * num;
+        long sq_num = num * num;
 
         String str_num = Integer.toString(num);
-        String squareOfNum = Integer.toString(sq_num);
+        String squareOfNum = Long.toString(sq_num);
 
         if (squareOfNum.endsWith(str_num)) {
             System.out.format("Number %d is Automorphic Number.", num);
